+++ src/components/ImageView/ImagePanel/ImagePanelComponent.tsx
@@ -6,7 +6,7 @@ import {observer} from "mobx-react";
 import {ImageViewLayer} from "components";
 import {CursorInfo, CursorInfoVisibility, ImageItem, ImageType, Zoom} from "models";
 import {PreviewWebGLService, TileService, TileWebGLService} from "services";
-import {AnimationMode, AppStore, type FrameStore} from "stores";
+import {AnimationMode, AppStore, ColorBlendingStore, type FrameStore} from "stores";
 
 import {BeamProfileOverlayComponent} from "../BeamProfileOverlay/BeamProfileOverlayComponent";
 import {CatalogViewGLComponent} from "../CatalogView/CatalogViewGLComponent";
@@ -157,9 +157,13 @@ export class ImagePanelComponent extends React.Component<ImagePanelComponentProp
                 }
             }
 
+            const showRaster = !isColorBlending || (isColorBlending && (this.props.image.store as ColorBlendingStore).rasterVisible);
+            const showContour = !isColorBlending || (isColorBlending && (this.props.image.store as ColorBlendingStore).contourVisible);
+            const showVector = !isColorBlending || (isColorBlending && (this.props.image.store as ColorBlendingStore).vectorOverlayVisible);
+
             return (
                 <div id={`image-panel-${this.props.column}-${this.props.row}`} className={className} style={style} onWheel={this.onMouseWheel} onMouseDown={this.onMouseDown} onMouseOver={this.onMouseEnter} onMouseLeave={this.onMouseLeave}>
-                    <RasterViewComponent
+                    {showRaster && <RasterViewComponent
                         image={this.props.image}
                         overlayStore={overlayStore}
                         webGLService={frame.isPreview ? PreviewWebGLService.Instance : TileWebGLService.Instance}
@@ -170,11 +174,11 @@ export class ImagePanelComponent extends React.Component<ImagePanelComponentProp
                         column={this.props.column}
                         tileBasedRender={!frame.isPreview}
                         rasterData={frame.previewPVRasterData}
-                    />
-                    <ContourViewComponent frame={frame} docked={this.props.docked} row={this.props.row} column={this.props.column} />
-                    <VectorOverlayViewComponent frame={frame} docked={this.props.docked} row={this.props.row} column={this.props.column} />
+                    />}
+                    {showContour && <ContourViewComponent frame={frame} docked={this.props.docked} row={this.props.row} column={this.props.column} />}
+                    {showVector && <VectorOverlayViewComponent frame={frame} docked={this.props.docked} row={this.props.row} column={this.props.column} />}
                     {frame?.overlayStore?.visible && <OverlayComponent image={this.props.image} overlaySettings={overlayStore} docked={this.props.docked} />}
-                    {this.cursorInfoRequired && frame.cursorInfo && (
+                    {this.cursorInfoRequired && this.frame.cursorInfo && !isColorBlending && (
                         <CursorOverlayComponent
                             cursorInfo={frame.cursorInfo}
                             cursorValue={frame.cursorInfo.isInsideImage ? (frame.isPreview ? frame.previewCursorValue.value : frame.cursorValue.value) : undefined}